+++ fidget/benches/render.rs
@@ -17,7 +17,7 @@ pub fn prospero_size_sweep(c: &mut Criterion) {
     for size in [256, 512, 768, 1024, 1280, 1546, 1792, 2048] {
         let cfg = &fidget::render::RenderConfig {
             image_size: size,
-            tile_sizes: fidget::vm::VmFunction::tile_sizes_2d().to_vec(),
+            tile_sizes: fidget::vm::VmFunction::tile_sizes_2d(),
             ..Default::default()
         };
         group.bench_function(BenchmarkId::new("vm", size), move |b| {
@@ -34,7 +34,7 @@ pub fn prospero_size_sweep(c: &mut Criterion) {
         {
             let cfg = &fidget::render::RenderConfig {
                 image_size: size,
-                tile_sizes: fidget::jit::JitFunction::tile_sizes_2d().to_vec(),
+                tile_sizes: fidget::jit::JitFunction::tile_sizes_2d(),
                 ..Default::default()
             };
             group.bench_function(BenchmarkId::new("jit", size), move |b| {
@@ -62,7 +62,7 @@ pub fn prospero_thread_sweep(c: &mut Criterion) {
     for threads in [1, 2, 4, 8, 16] {
         let cfg = &fidget::render::RenderConfig {
             image_size: 1024,
-            tile_sizes: fidget::vm::VmFunction::tile_sizes_2d().to_vec(),
+            tile_sizes: fidget::vm::VmFunction::tile_sizes_2d(),
             threads: threads.try_into().unwrap(),
             ..Default::default()
         };
@@ -79,7 +79,7 @@ pub fn prospero_thread_sweep(c: &mut Criterion) {
         {
             let cfg = &fidget::render::RenderConfig {
                 image_size: 1024,
-                tile_sizes: fidget::jit::JitFunction::tile_sizes_2d().to_vec(),
+                tile_sizes: fidget::jit::JitFunction::tile_sizes_2d(),
                 threads: threads.try_into().unwrap(),
                 ..Default::default()
             };